renaming the module to be more representative
